LayoutTests/media/media-source/media-source-detach-removes-all-text-tracks-expected.txt


EXPECTED (source.detachable == 'true') OK
RUN(video.srcObject = source)
EVENT(sourceopen)
EVENT(updateend)
EXPECTED (video.textTracks.length == '2') OK
RUN(video.srcObject = null)
EXPECTED (removetrackCount == '2') OK
END OF TEST
